requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
pdfminer.six
python-multipart
